Added support for revocation reasons (other than 'unspecified').
Increased the expiry period of generated CRLs to 10 years, so strongswan does not unexpectedly stop working,
when the CRL of the internal CA expires. If a client certificate is revoked, strongswan must be restarted for
the change to become active. A better solution would be to use an OCSP responder (overkill for the internal CA).
Nevertheless, for larger setups with many clients an external CA supporting OCSP is recommended.